Entregable_4/airflow/scripts/entregable2/requirements/requirements.txt
requests
redshift-connector
python-dotenv
numpy
pandas
psycopg2-binary
sqlalchemy